requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
websockets>=11.0
cryptography>=40.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
eth-account>=0.8.0
web3>=6.0.0
flask>=2.3.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
